Fix for Issue 960: Result.getAsInteger(path) throws number format exception

+++ CodenameOne/src/com/codename1/processing/Result.java
@@ -385,7 +385,7 @@ public class Result {
 		if (s == null) {
 			return 0;
 		}
-		return Integer.parseInt(s);
+		return (int)Double.parseDouble(s);
 	}
 
 	/**
